requirements.txt
transformers
python-dotenv
requests